+++ CHANGELOG.md
@@ -8,11 +8,13 @@
 
 ### Changed
 
-- Improve menu title of tools
+- Improve title of tools in the menu
 
 ### Removed
 
 ### Fixed
+
+- Copy action in the Password Generator tool contained escaped HTML entities
 
 ## [6.0.1] - 2024-11-22