+++ CHANGELOG.md
@@ -4,19 +4,27 @@
 
 ### Added
 
+### Changed
+
+### Removed
+
+### Fixed
+
+## 5.2.0 - 2024-12-13
+
+### Added
+
 - Add ASCII Encoder/Decoder tool
 
 ### Changed
 
 - Improve title of tools in the menu
 
-### Removed
-
 ### Fixed
 
 - Copy action in the Password Generator tool contained escaped HTML entities
 
-## [5.1.1] - 2024-11-22
+## 5.1.1 - 2024-11-22
 
 ### Fixed